// AppCovid-19/src/models/usuario.model.ts
export class UsuarioModel {
    idUsuario: number;
    idInstancia: number;
    idTipoUsuario: number;
    idGenero: number;
    nombre: string;
    apellidoP: string;
    apellidoM: string;
    nacionalidad: string;
    anioNacimiento: number;
    calle: string;
    numero: number;
    municipio: string;
    codigoPostal: string;
    telefono: string;
    email: string;
    contrasenia: string;
    activo: boolean;
    token: string;
    recibe_notificacion: boolean;
    descripcionInstancia: string;
    descripcionTUsuario: string;
    descripcionGenero: string;
    idDependencia: number;
}