// AppCovid-19/src/models/instancia.model.ts
export class InstanciaModel {
    idInstancia: number;
    descripcion: string;
    activo: boolean;
    nombreCorto: string;
    idEstado: number;
    idMunicipio: number;
    nombreUsuario: string;
    apellidoP: string;
    apellidoM: string;
    email: string;
    password: string;
    descripcionEstado: string;
    descripcionMunicipio: string;
    logo: string;
    poligono_json: string;
    idEncuesta: number;
}